Ported whole engine, gamelogic and tools to C++

+++ tools/bspc/l_mem.h
@@ -45,7 +45,7 @@
 
 #ifndef MEMDEBUG
 
-void* GetClearedMemory( int size );
+void* GetClearedMemory(unsigned long size);
 void* GetMemory( unsigned long size );
 
 #else
